[client-app] Fix version conflict with send-reminders
Summary:
The 409 version conflict error was due to an actual version conflict
when we were syncing back metadata. When you set a reminder on a draft
that hasn't been sent yet, we initially set just the expiration date in
the metadata. Then, when the draft is actually sent, we update the
metadata with some fields that aren't finalized until after the send
(e.g. messageIdHeader, subject, replyTo). However, this second update
was being made before we received the delta for the first update, so the
version number in the second update was incorrect.

To fix this, this diff adds some special handling in the metadata route
to detect if the metadata version is off by one. If data was only added,
not changed, then we merge the two versions together.

Test Plan: manual, specs

Reviewers: juan, spang, evan

Reviewed By: evan

Differential Revision: https://phab.nylas.com/D4383