[client-sync] fix sending when racing syncing of Gmail sent folder
Summary:
On gmail if you sent a message to multiple people with big attachments,
the send could take several minutes.

In the meantime, some of your individually sent messages would show up in
your Gmail's sent folder.

When the send completed, we'd try and create a new message object with the
same ID and fail.

This changes it to upsert instead

TODO to see how this behaves with replies

Test Plan: manual

Reviewers: juan, halla

Reviewed By: halla

Differential Revision: https://phab.nylas.com/D4490